+++ src/main/java/org/kakaoshare/backend/domain/gift/repository/query/GiftRepositoryCustom.java
@@ -1,8 +1,8 @@
 package org.kakaoshare.backend.domain.gift.repository.query;
 
-import org.kakaoshare.backend.domain.gift.dto.GiftDescriptionResponse;
-import org.kakaoshare.backend.domain.gift.dto.GiftDetailResponse;
-import org.kakaoshare.backend.domain.gift.dto.GiftDto;
+import org.kakaoshare.backend.domain.gift.dto.inquiry.detail.response.GiftDescriptionResponse;
+import org.kakaoshare.backend.domain.gift.dto.inquiry.detail.response.GiftDetailResponse;
+import org.kakaoshare.backend.domain.gift.dto.inquiry.history.GiftDto;
 import org.kakaoshare.backend.domain.gift.entity.GiftStatus;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
