+++ src/main/java/org/kakaoshare/backend/domain/gift/repository/query/GiftRepositoryCustomImpl.java
@@ -3,10 +3,10 @@ package org.kakaoshare.backend.domain.gift.repository.query;
 import com.querydsl.jpa.impl.JPAQuery;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import lombok.RequiredArgsConstructor;
-import org.kakaoshare.backend.domain.gift.dto.GiftDescriptionResponse;
-import org.kakaoshare.backend.domain.gift.dto.GiftDetailResponse;
-import org.kakaoshare.backend.domain.gift.dto.GiftDto;
-import org.kakaoshare.backend.domain.gift.dto.QGiftDto;
+import org.kakaoshare.backend.domain.gift.dto.inquiry.detail.response.GiftDescriptionResponse;
+import org.kakaoshare.backend.domain.gift.dto.inquiry.detail.response.GiftDetailResponse;
+import org.kakaoshare.backend.domain.gift.dto.inquiry.history.GiftDto;
+import org.kakaoshare.backend.domain.gift.dto.inquiry.history.QGiftDto;
 import org.kakaoshare.backend.domain.gift.entity.Gift;
 import org.kakaoshare.backend.domain.gift.entity.GiftStatus;
 import org.kakaoshare.backend.domain.gift.exception.GiftErrorCode;
