+++ src/main/java/org/kakaoshare/backend/domain/gift/service/GiftService.java
@@ -3,9 +3,9 @@ package org.kakaoshare.backend.domain.gift.service;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.kakaoshare.backend.common.dto.PageResponse;
-import org.kakaoshare.backend.domain.gift.dto.GiftDescriptionResponse;
-import org.kakaoshare.backend.domain.gift.dto.GiftDetailResponse;
-import org.kakaoshare.backend.domain.gift.dto.GiftResponse;
+import org.kakaoshare.backend.domain.gift.dto.inquiry.detail.response.GiftDescriptionResponse;
+import org.kakaoshare.backend.domain.gift.dto.inquiry.detail.response.GiftDetailResponse;
+import org.kakaoshare.backend.domain.gift.dto.inquiry.history.response.GiftResponse;
 import org.kakaoshare.backend.domain.gift.entity.GiftStatus;
 import org.kakaoshare.backend.domain.gift.repository.GiftRepository;
 import org.kakaoshare.backend.domain.member.repository.MemberRepository;
