+++ src/test/java/org/kakaoshare/backend/domain/gift/service/GiftServiceTest.java
@@ -7,10 +7,10 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.EnumSource;
 import org.kakaoshare.backend.common.dto.PageResponse;
-import org.kakaoshare.backend.domain.gift.dto.GiftDescriptionResponse;
-import org.kakaoshare.backend.domain.gift.dto.GiftDetailResponse;
-import org.kakaoshare.backend.domain.gift.dto.GiftDto;
-import org.kakaoshare.backend.domain.gift.dto.GiftResponse;
+import org.kakaoshare.backend.domain.gift.dto.inquiry.detail.response.GiftDescriptionResponse;
+import org.kakaoshare.backend.domain.gift.dto.inquiry.detail.response.GiftDetailResponse;
+import org.kakaoshare.backend.domain.gift.dto.inquiry.history.GiftDto;
+import org.kakaoshare.backend.domain.gift.dto.inquiry.history.response.GiftResponse;
 import org.kakaoshare.backend.domain.gift.entity.GiftStatus;
 import org.kakaoshare.backend.domain.gift.repository.GiftRepository;
 import org.kakaoshare.backend.domain.member.entity.Member;
